+++ changelog.md
@@ -1,5 +1,6 @@
 ### v1.0.3 (Sept. 15, 2019)
 * Added Darkshore NPCs
+* Added Redridge Mountains NPCs
 * Minor NPC fixes in Stormwind and Darnassus
 
 [See past changes here.](https://bitbucket.org/jsiebert9/townsfolk-tracker/src/master/changehistory.md)